% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_na.R
\name{delete_na}
\alias{delete_na}
\title{Remove NA from list}
\usage{
delete_na(y = list(), recursive = FALSE)
}
\arguments{
\item{y}{a list}

\item{recursive}{logical. If \code{TRUE}, \code{NA} within the list's
elements are removed too.}
}
\value{
Returns the list without \code{NA}.
}
\description{
This functions removes \code{NA} from a list. With \code{recursive = TRUE},
\code{NA} within each list's elements are removed as well.
}
\examples{
y <- list(c(1:3), letters[1:4], NA, c(1, NA), list(c(5:6, NA), NA, "A"))
delete_na(y, recursive = TRUE)

# [[1]]
# [1] 1 2 3
#
# [[2]]
# [1] "a" "b" "c" "d"
#
# [[3]]
# [1] 1
#
# [[4]]
# [[4]][[1]]
# [1] 5 6
#
# [[4]][[2]]
# [1] "A"
}
