% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit.R
\name{StrSplitP}
\alias{StrSplitP}
\title{The improvement of \code{strsplit}}
\usage{
StrSplitP(x, split, type = "remove", perl = FALSE, ...)
}
\arguments{
\item{x}{character vector, each element of which is to be split. Other
inputs, including a factor, will give an error.}

\item{split}{character vector (or object which can be coerced to such)
containing regular expression(s) to use for splitting. If empty matches
occur, in particular if \code{split} has length 0, \code{x} is split into single
characters. If \code{split} has length greater than 1, it is recycled along \code{x}.}

\item{type}{a character. Either to \code{remove} or keep the delimiter
\code{before}, \code{between }or \code{remove} the split.}

\item{perl}{logical.}

\item{...}{other inputs passed along to \code{strsplit}.}
}
\value{
A list of the same length as \code{x}, the i-th element of which contains
the vector
}
\description{
This functions uses \code{strsplit} and adds the possibility to split
and keep the delimiter after or before the given split. Or it can split
between two given splits and keep both.
}
\examples{
x <- c("3D/MON&SUN")
StrSplitP(x, "", type = "before")
x <- c("3D/MON&SUN")
}
